requirements.txt
dnspython==2.6.1
aiohttp==3.9.5
asyncio-throttle==1.0.2
rich==13.7.1
